moved all node actions to a specific node, and created a component for node rows
