+++ src/bin/stack/app/src/nodes/AddNode.svelte
@@ -7,9 +7,9 @@
     TextInput,
   } from "carbon-components-svelte";
   import Add from "carbon-icons-svelte/lib/Add.svelte";
-  import { allNodeTypes } from "./nodes";
-  import type { NodeType, Node } from "./nodes";
-  import { stack } from "./store";
+  import { allNodeTypes } from "../nodes";
+  import type { NodeType, Node } from "../nodes";
+  import { stack } from "../store";
 
   let open = false;
   let name = "";
